backend/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn==0.32.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & Security
PyJWT==2.10.1
bcrypt==4.2.1
cryptography==43.0.3

# Real-time Communication
python-socketio==5.12.0
python-multipart==0.0.12
python-engineio==4.11.0

# Caching & Messaging
redis==5.2.1

# Data Validation
pydantic==2.10.3
pydantic-settings==2.7.0

# Environment Variables
python-dotenv==1.0.1

# CORS
fastapi-cors==0.0.6

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Optional: Better error tracking
sentry-sdk==2.20.0

# Optional: Better logging
python-json-logger==2.0.7